Sum of Numbers from TXT file
Write a program to read a text file "Input.txt" holding a sequence of
integer numbers, each at a separate line. Print the sum of the numbers
at the console. Ensure you close correctly the file in case of exception
or in case of normal execution. In case of exception (e.g. the file is
missing), print "Error" as a result.